requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
bs4
langchain
cassio
beautifulsoup4
langchainhub
google-generativeai
streamlit
pytesseract
pillow
requests
opencv-python
ipykernel
langchain_huggingface
pypdf
unstructured
pdfminer
chromadb
rank_bm25
langgraph
langsmith
langchain
langchain_groq
langchain_community
